requirements.txt
faiss-cpu
sentence-transformers
python-dotenv
fastapi
uvicorn
openai
